ARM: OMAP: mailbox: optimize save and restore routines
Current save and restore context functions iterate on a bunch
of predefined registers, most of them either are read only or
doesn't matter for context saving purposes.

From mailbox perspective, the important registers to be restored
are the irqenable and the messages (if any) left in the HW queue.
Until now unprocessed messages in the HW queue were ignored and
lost. Mailbox needs to take care of the messages left on its TX
queue, this offset has now been defined.

Irqdisable register is ignored because it is a mirror of irqenable,
in OMAP3 both are the same register, and in OMAP4 both contain the
same value, so for context save we don't care to save a disabled irq.

Functions: omap2_mbox_disable_irq and omap2_mbox_enable_irq will
store the updated value to restore for irqenable, for this reason
a variant of these function is needed when saving and restoring that
doesn't update this value, these are internal and should never be
interfaced with the common framework.

Signed-off-by: Omar Ramirez Luna <omar.luna@linaro@.org>